Manage method functional implementation

+++ app/Http/Controllers/ManagementController.php
@@ -10,10 +10,13 @@ use App\Http\Controllers\UserController;
 
 class ManagementController extends Controller {
 
-    public function showAnAttachForm(Fridge $fridge, User $user) {
+    public function showAManageForm(Fridge $fridge) {
         if(Auth::user()->isFridgeOwner($fridge))
         {
-            return view('management.attach', compact('fridge', 'user'));
+            return view('management.manage', [
+                'fridge' => $fridge,
+                'users' => $fridge->users()->get(),
+            ]);
         }
         else
         {
@@ -21,24 +24,25 @@ class ManagementController extends Controller {
         }
     }
 
-    public function attachUserToFridge(Fridge $fridge, User $user, Request $request)
+    public function attachUserToFridge(Fridge $fridge, Request $request)
     {
 
         if(Auth::user()->isFridgeOwner($fridge))
         {
             $request->validate([
                 'is_owner' => 'required|numeric|min:0|max:1',
+                'user_id' => 'required|numeric|exists:users,id',
             ]);
-            $fridge->users()->attach($user->id, ['is_owner' => $request->is_owner]);
+            $fridge->users()->attach($request->user_id, ['is_owner' => $request->is_owner]);
             return redirect()->route('myfridges.indexOwn');
         } else {
             abort(403, 'Access denied');
         }
     }
 
-    public function detachUserFromFridge(Fridge $fridge, User $user, Request $request)
+    public function detachUserFromFridge(Fridge $fridge, User $user)
     {
-        if(Auth::user()->isFridgeOwner($fridge))
+        if(Auth::user()->isFridgeOwner($fridge) && $user->isFridgeUser($fridge))
         {
             $fridge->users()->detach($user->id);
             return redirect()->route('myfridges.indexOwn');
@@ -51,9 +55,7 @@ class ManagementController extends Controller {
     {
         if(Auth::user()->isFridgeUserNoOwner($fridge))
         {
-            $user = Auth::user();
-            $fridge->users()->detach($user->id);
-            $user->fridges()->detach($fridge->id);
+            $fridge->users()->detach(Auth::user()->id);
             return redirect()->route('myfridges.indexOwn');
         } else {
             abort(403, 'Access denied');
